pyproject.toml
[project]
name = "basicscore"
description = "Creates basicness-scores of software"
readme = "Readme.md"
version = "0.0.2"
authors = [
    { name = "Caj Larsson", email = "basic@caj.me" },
    { name= "Billy Bonks", email = "no@example.org" }
]
dependencies = [
    "tqdm",
    "llama-cpp-python",
    "appdirs"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
basicscore = "basicscore:cli.main"

[tool.hatch.build]
include = ["basicscore"]

[project.optional-dependencies]
dev = [
    "ipython"
]
